refactor unit test using ResultAssertUtils.
