+++ sharding-jdbc-example/raw-jdbc-nodep-example/src/test/java/io/shardingsphere/example/jdbc/nodep/YamlConfigurationExampleTest.java
@@ -19,47 +19,48 @@ package io.shardingsphere.example.jdbc.nodep;
 
 import io.shardingsphere.example.jdbc.nodep.factory.YamlCommonServiceFactory;
 import io.shardingsphere.example.repository.api.senario.CommonServiceScenario;
+import io.shardingsphere.example.repository.api.trace.ResultAssertUtils;
 import io.shardingsphere.example.type.ShardingType;
 import org.junit.Test;
 
 import java.io.IOException;
 import java.sql.SQLException;
 
-public class YamlConfigurationExampleTest extends BaseConfigurationExample {
+public class YamlConfigurationExampleTest {
     
     @Test
     public void assertShardingDatabasePrecise() throws SQLException, IOException {
         CommonServiceScenario scenario = new CommonServiceScenario(YamlCommonServiceFactory.newInstance(ShardingType.SHARDING_DATABASES));
         scenario.executeShardingCRUDSuccess();
-        assertShardingDatabaseResult(scenario.getCommonService(), false);
+        ResultAssertUtils.assertShardingDatabaseResult(scenario.getCommonService(), false);
     }
     
     @Test
     public void assertShardingTablesPrecise() throws SQLException, IOException {
         CommonServiceScenario scenario = new CommonServiceScenario(YamlCommonServiceFactory.newInstance(ShardingType.SHARDING_TABLES));
         scenario.executeShardingCRUDSuccess();
-        assertShardingTableResult(scenario.getCommonService(), false);
+        ResultAssertUtils.assertShardingTableResult(scenario.getCommonService(), false);
     }
     
     @Test
     public void assertShardingDatabaseAndTablesPrecise() throws SQLException, IOException {
         CommonServiceScenario scenario = new CommonServiceScenario(YamlCommonServiceFactory.newInstance(ShardingType.SHARDING_DATABASES_AND_TABLES));
         scenario.executeShardingCRUDSuccess();
-        assertShardingDatabaseAndTableResult(scenario.getCommonService(), false);
+        ResultAssertUtils.assertShardingDatabaseAndTableResult(scenario.getCommonService(), false);
     }
     
     @Test
     public void assertMasterSlave() throws SQLException, IOException {
         CommonServiceScenario scenario = new CommonServiceScenario(YamlCommonServiceFactory.newInstance(ShardingType.MASTER_SLAVE));
         scenario.executeShardingCRUDSuccess();
-        assertMasterSlaveResult(scenario.getCommonService());
+        ResultAssertUtils.assertMasterSlaveResult(scenario.getCommonService());
     }
     
     @Test
     public void assertShardingMasterSlavePrecise() throws SQLException, IOException {
         CommonServiceScenario scenario = new CommonServiceScenario(YamlCommonServiceFactory.newInstance(ShardingType.SHARDING_MASTER_SLAVE));
         scenario.executeShardingCRUDSuccess();
-        assertMasterSlaveResult(scenario.getCommonService());
+        ResultAssertUtils.assertMasterSlaveResult(scenario.getCommonService());
     }
 }
